Cuepoints-manager version 2.0
It's developed and debugged in tested for Ableton Live 11.3.21 and Ableton Live 12.0.25 both with max 8.6.5

Numerous changes have been made to menu management to reduce the device's interference with the Undo History.
Network interactions have been improved, and several bugs have been fixed.
The data structure of the cues has been updated to allow duplicate names between locators and names with spaces.

IMPORTANT:
To update your project to the latest version, follow these steps:

Replace the device with the current version.
Save the Ableton session with a new name and reopen it.
Click on the "data" label, then double-click the object [dict dict_cues @quiet 1 @legacy 1].
Delete all the content except for the two curly braces { }.

WARNING: This operation deletes the active cues in the project, which will need to be re-entered manually. It is recommended to document the original configuration beforehand.